+++ springboot_example/src/main/java/com/github/braisdom/example/model/Product.java
@@ -7,13 +7,9 @@ import com.github.braisdom.objsql.annotations.Queryable;
 public class Product {
     @Queryable
     private String barcode;
-
     @Queryable
     private String name;
-
     private Integer categoryId;
-
-    private Double salesPrice;
-
-    private Double cost;
+    private Float salesPrice;
+    private Float cost;
 }
